LPS-124961 Move *ConfigurationPidMapping to *.internal.configuration.definition
